alibabacloud: Rename eni package to ipam
Rename `pkg/alibabacloud/eni` to `pkg/alibabacloud/ipam` so the
operator-side IPAM allocator (the `ipamTypes.NodeOperations`
implementation that runs in the Cilium Operator) lives under a name
that describes its role rather than the AWS-borrowed "eni" label,
matching a standard cloud provider package layout of
`pkg/$cloud/{api,metadata,types,ipam}`.
