+++ pkg/alibabacloud/ipam/log.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package eni
+package ipam
 
 import (
 	"github.com/cilium/cilium/pkg/logging/logfields"
